+++ context/application.go
@@ -48,4 +48,8 @@ type Application interface {
 	// If a handler is not already registered,
 	// then it creates & registers a new trivial handler on the-fly.
 	FireErrorCode(ctx Context)
+
+	// RouteExists reports whether a particular route exists
+	// It will search from the current subdomain of context's host, if not inside the root domain.
+	RouteExists(ctx Context, method, path string) bool
 }